create `googleOAuth.ts` and fix Back to Login arrow
